Move self-signed certificates to test-containers folder to escape from accidental security issues
Before creation of the container, we always run our tests and if we keep
test only certificate on top level of the repository, maybe we can
accidently keep it part of container context and it can create a bigger
security problem. I just moved them to test-containers folder and
ignored all certificate files from the docker context.