Auth Ldap SSL (#136)
* add ldaps with ca auth

* ldap certs

* logging ssl issue

* remove pyssl of dev dependencies

* ssl cert generating

* ssl generate

* optimize

* ssl working

* ssl

* optimize

* ssl gen works

* args

* docs

* move

* ssl test with cets

* fix

* Move self-signed certificates to test-containers folder to escape from accidental security issues

Before creation of the container, we always run our tests and if we keep
test only certificate on top level of the repository, maybe we can
accidently keep it part of container context and it can create a bigger
security problem. I just moved them to test-containers folder and
ignored all certificate files from the docker context.

---------

Co-authored-by: Philipp Kraus <philipp.kraus@flashpixx.de>
Co-authored-by: Burak Ince <burak.ince@linux.org.tr>